ported groups

+++ competition-serv-model/src/main/java/compman/compsrv/model/dto/brackets/CompetitorStageResultDTO.java
@@ -5,15 +5,13 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Accessors(chain = true)
 public class CompetitorStageResultDTO {
     private String competitorId;
-    private BigDecimal points;
+    private Integer points;
     private Integer round;
     private StageRoundType roundType;
     private Integer place;
